+++ src/auth/dto/response/index.ts
@@ -1,3 +1,4 @@
 export * from './auth-tokens.dto';
 export * from './user-public.dto';
 export * from './auth-response.dto';
+export * from './email-response.dto';